+++ lib/src/extensions/nonet_extensions.dart
@@ -121,34 +121,26 @@ extension NonetExtensions<T0, T1, T2, T3, T4, T5, T6, T7, T8> on Nonet<T0, T1, T
   ///
   /// - Note: If the [Nonet] does not contain a value of type [R], the [callback] is not executed.
   void ifType<R>({
-    required void Function(T0) first,
-    required void Function(T1) second,
-    required void Function(T2) third,
-    required void Function(T3) fourth,
-    required void Function(T4) fifth,
-    required void Function(T5) sixth,
-    required void Function(T6) seventh,
-    required void Function(T7) eighth,
-    required void Function(T8) ninth,
+    required void Function(R) callback,
   }) {
-    if (isFirst) {
-      return first(this.first);
-    } else if (isSecond) {
-      return second(this.second);
-    } else if (isThird) {
-      return third(this.third);
-    } else if (isFourth) {
-      return fourth(this.fourth);
-    } else if (isFifth) {
-      return fifth(this.fifth);
-    } else if (isSixth) {
-      return sixth(this.sixth);
-    } else if (isSeventh) {
-      return seventh(this.seventh);
-    } else if (isEighth) {
-      return eighth(this.eighth);
-    } else if (isNinth) {
-      return ninth(this.ninth);
+    if (isFirst && first is R) {
+      callback(first as R);
+    } else if (isSecond && second is R) {
+      callback(second as R);
+    } else if (isThird && third is R) {
+      callback(third as R);
+    } else if (isFourth && fourth is R) {
+      callback(fourth as R);
+    } else if (isFifth && fifth is R) {
+      callback(fifth as R);
+    } else if (isSixth && sixth is R) {
+      callback(sixth as R);
+    } else if (isSeventh && seventh is R) {
+      callback(seventh as R);
+    } else if (isEighth && eighth is R) {
+      callback(eighth as R);
+    } else if (isNinth && ninth is R) {
+      callback(ninth as R);
     }
   }
 }