+++ lib/src/types/unauthorized.dart
@@ -1,0 +1,8 @@
+import 'package:any_of/any_of.dart';
+
+class Unauthorized extends AnyOfType {
+  Unauthorized();
+
+  @override
+  String toString() => 'Unauthorized';
+}